+++ components/faphub/installedtab/impl/src/main/java/com/flipperdevices/faphub/installedtab/impl/viewmodel/InstalledFapsFromNetworkProducer.kt
@@ -31,14 +31,17 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import javax.inject.Inject
+import javax.inject.Singleton
 
+@Singleton
 class InstalledFapsFromNetworkProducer @Inject constructor(
     private val installedFapsUidsProducer: InstalledFapsUidsProducer,
     private val flipperTargetProviderApi: FlipperTargetProviderApi,
     private val fapNetworkApi: FapNetworkApi,
     private val fapStateManager: FapInstallationStateManager,
+    private val globalScope: CoroutineScope
 ) : LogTagProvider {
-    override val TAG = "InstalledFapsFromNetworkProducer"
+    override val TAG = "InstalledFapsFromNetworkProducer-${hashCode()}"
 
     private val applicationFromNetworkStateFlow = MutableStateFlow<FapInstalledFromNetworkState>(
         FapInstalledFromNetworkState.Loading
@@ -46,12 +49,15 @@ class InstalledFapsFromNetworkProducer @Inject constructor(
     private var fetchFromNetworkJob: Job? = null
     private val mutex = Mutex()
 
-    fun refresh(scope: CoroutineScope, force: Boolean) {
-        installedFapsUidsProducer.refresh(scope, force)
+    fun refresh(force: Boolean) {
+        info { "Call refresh $force" }
+        installedFapsUidsProducer.refresh(globalScope, force)
         val oldJob = fetchFromNetworkJob
-        fetchFromNetworkJob = scope.launch(Dispatchers.Default) {
+        fetchFromNetworkJob = globalScope.launch(Dispatchers.Default) {
             oldJob?.cancelAndJoin()
-            applicationFromNetworkStateFlow.emit(FapInstalledFromNetworkState.Loading)
+            if (force) {
+                applicationFromNetworkStateFlow.emit(FapInstalledFromNetworkState.Loading)
+            }
             combine(
                 installedFapsUidsProducer.getUidsStateFlow(),
                 flipperTargetProviderApi.getFlipperTarget().filterNotNull()
@@ -76,6 +82,7 @@ class InstalledFapsFromNetworkProducer @Inject constructor(
                             FapInstalledFromNetworkState.Error(throwable)
                         }
                     }
+                    info { "Update applicationFromNetworkStateFlow with state $state" }
                     applicationFromNetworkStateFlow.emit(state)
                 }
             }
@@ -145,13 +152,16 @@ class InstalledFapsFromNetworkProducer @Inject constructor(
         flipperTarget: FlipperTarget
     ): FapInstalledFromNetworkState {
         val toRequestItems = uids.toMutableSet()
-        val alreadyLoadedApps = when (val currentState = applicationFromNetworkStateFlow.value) {
+        val currentLoadedItems = when (val currentState = applicationFromNetworkStateFlow.value) {
             is FapInstalledFromNetworkState.Error,
             is FapInstalledFromNetworkState.LoadedOffline,
             FapInstalledFromNetworkState.Loading -> persistentListOf()
 
             is FapInstalledFromNetworkState.Loaded -> currentState.faps
-        }.filter { (fapTarget, _) -> fapTarget == flipperTarget }
+        }
+        info { "currentLoadedItems: ${currentLoadedItems.map { it.second.id }}" }
+        val alreadyLoadedApps = currentLoadedItems
+            .filter { (fapTarget, _) -> fapTarget == flipperTarget }
             .filter { (_, fapItem) -> toRequestItems.contains(fapItem.id) }
 
         alreadyLoadedApps.forEach { (_, fapItem) ->
@@ -188,7 +198,7 @@ class InstalledFapsFromNetworkProducer @Inject constructor(
 }
 
 private sealed class FapInstalledFromNetworkState {
-    object Loading : FapInstalledFromNetworkState()
+    data object Loading : FapInstalledFromNetworkState()
 
     data class LoadedOffline(
         val faps: ImmutableList<OfflineFapApp>