+++ components/hub/api/src/main/java/com/flipperdevices/hub/api/HubApi.kt
@@ -1,7 +1,8 @@
 package com.flipperdevices.hub.api
 
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 
 interface HubApi {
-    fun hasNotification(): Flow<Boolean>
+    fun hasNotification(scope: CoroutineScope): Flow<Boolean>
 }