+++ components/hub/impl/src/main/java/com/flipperdevices/hub/impl/api/HubApiImpl.kt
@@ -1,17 +1,27 @@
 package com.flipperdevices.hub.impl.api
 
 import com.flipperdevices.core.di.AppGraph
+import com.flipperdevices.faphub.installedtab.api.FapUpdatePendingCountApi
 import com.flipperdevices.hub.api.HubApi
 import com.flipperdevices.nfc.attack.api.NfcAttackApi
 import com.squareup.anvil.annotations.ContributesBinding
-import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.combine
 import javax.inject.Inject
+import javax.inject.Provider
 
 @ContributesBinding(AppGraph::class)
 class HubApiImpl @Inject constructor(
-    private val nfcAttackApi: NfcAttackApi
+    private val nfcAttackApi: Provider<NfcAttackApi>,
+    private val fapUpdatePendingCountApi: Provider<FapUpdatePendingCountApi>
 ) : HubApi {
-    override fun hasNotification() = nfcAttackApi.notificationCount().map {
-        it > 0
+    override fun hasNotification(scope: CoroutineScope): Flow<Boolean> {
+        return combine(
+            nfcAttackApi.get().notificationCount(),
+            fapUpdatePendingCountApi.get().getUpdatePendingCount()
+        ) { (nfcNotificationCount, appUpdates) ->
+            nfcNotificationCount > 0 || appUpdates > 0
+        }
     }
 }