+++ components/updater/impl/src/main/java/com/flipperdevices/updater/impl/UpdaterTask.kt
@@ -10,6 +10,7 @@ import com.flipperdevices.core.log.error
 import com.flipperdevices.core.log.info
 import com.flipperdevices.core.preference.FlipperStorageProvider
 import com.flipperdevices.core.ui.lifecycle.OneTimeExecutionBleTask
+import com.flipperdevices.faphub.installedtab.api.FapNeedUpdatePopUpHelper
 import com.flipperdevices.updater.impl.model.IntFlashFullException
 import com.flipperdevices.updater.impl.model.UpdateContentException
 import com.flipperdevices.updater.impl.tasks.FlipperUpdateImageHelper
@@ -34,13 +35,15 @@ import java.net.UnknownHostException
 
 private const val DISCONNECT_WAIT_TIMEOUT_MS = 30 * 1000L
 
+@Suppress("LongParameterList")
 class UpdaterTask(
     serviceProvider: FlipperServiceProvider,
     private val context: Context,
     private val uploadToFlipperHelper: UploadToFlipperHelper,
     private val subGhzProvisioningHelper: SubGhzProvisioningHelper,
     private val updateContentDownloader: MutableSet<UpdateContentDownloader>,
-    private val flipperStorageApi: FlipperStorageApi
+    private val flipperStorageApi: FlipperStorageApi,
+    private val fapNeedUpdatePopUpHelper: FapNeedUpdatePopUpHelper
 ) : OneTimeExecutionBleTask<UpdateRequest, UpdatingState>(serviceProvider),
     LogTagProvider {
     override val TAG = "UpdaterTask"
@@ -158,6 +161,9 @@ class UpdaterTask(
             serviceApi.connectionInformationApi.getConnectionStateFlow()
                 .filter { !it.isConnected }.first()
         }
+
+        fapNeedUpdatePopUpHelper.notifyIfUpdateAvailable()
+
         stateListener(UpdatingState.Rebooting)
     }